Delete utils/read.py and move its contents to main_parallelized.py
